Revert "fixed lang and theme changers"
