+++ src/components/Header/Header.jsx
@@ -1,38 +1,26 @@
 import { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import ThemeToggler from '../ThemeToggler/ThemeToggler';
 import Container from '../Container/Container';
 import Logo from '../Logo/Logo';
 import UserAuth from '../UserAuth/UserAuth';
 import UserLogo from '../UserLogo/UserLogo';
 import LanguageSelector from '../LanguageSelector/LanguageSelector';
-import { selectColorTheme, selectIsLogged } from '../../redux/selectors';
+import { selectIsLogged } from '../../redux/selectors';
 import { selectLanguage } from '../../redux/Global/selectors';
 import i18n, { LANGUAGES } from '../../Internationalization/i18n';
 import { setLanguage } from '../../redux/Global/globalSlice';
 import { LngLogoutWrap, StyledBox, StyledWrapper } from './Header.styled';
-import { changeUserTheme } from '../../redux/auth/authSlice';
-import { updateUserThunk } from '../../redux/auth/operations';
 
 export const Header = () => {
   const dispatch = useDispatch();
   const isAuthenticated = useSelector(selectIsLogged);
   const language = useSelector(selectLanguage);
-  const theme = useSelector(selectColorTheme);
   const [languageActive, setLanguageActive] = useState(false);
 
   useEffect(() => {
     setLanguageActive(language === LANGUAGES.EN);
     i18n.changeLanguage(language);
   }, [language]);
-
-  const changeTheme = () => {
-    const newTheme = theme === 'light' ? 'dark' : 'light';
-    document.documentElement.classList.remove(theme);
-    document.documentElement.classList.add(newTheme);
-    dispatch(changeUserTheme(newTheme));
-    dispatch(updateUserThunk({ theme: newTheme }));
-  };
 
   const onLanguageChange = (e) => {
     let selectedLanguage;
@@ -45,20 +33,16 @@ export const Header = () => {
     }
     dispatch(setLanguage(selectedLanguage));
   };
-
   return (
     <header>
       <Container>
         <StyledWrapper>
           <Logo />
-          <div>
+          <LngLogoutWrap>
             <LanguageSelector
               checked={languageActive}
               onChange={onLanguageChange}
             />
-            {isAuthenticated && <ThemeToggler onClick={changeTheme} />}
-          </div>
-          <LngLogoutWrap>
             <StyledBox>
               {isAuthenticated ? <UserLogo /> : <UserAuth />}
             </StyledBox>
