+++ src/components/SettingModal/SettingModal.jsx
@@ -10,11 +10,12 @@ import '../../Internationalization/i18n';
 import sprite from '../../assets/sprite.svg';
 import defaultPhoto from '../../assets/avatar.jpg';
 import EyePassButton from './EyePassBtn';
+import ThemeToggler from '../ThemeToggler/ThemeToggler';
 import {
   updateAvatarThunk,
   updateUserThunk,
 } from '../../redux/auth/operations';
-import { selectUser } from '../../redux/selectors';
+import { selectColorTheme, selectUser } from '../../redux/selectors';
 import { changeModalOpen, closeModals } from '../../redux/modals/modalsSlice';
 import {
   LabelText,
@@ -32,11 +33,13 @@ import {
   StyledImg,
   TitleSwitcherWrap,
 } from './SettingModal.styled';
+import { changeUserTheme } from '../../redux/auth/authSlice';
 
 const SettingModal = () => {
   const [eyePass, setEyePass] = useState(false);
   const { t } = useTranslation();
   const { avatarURL, gender, name, email } = useSelector(selectUser);
+  const theme = useSelector(selectColorTheme);
 
   const dispatch = useDispatch();
   const basicSchema = yup
@@ -138,10 +141,19 @@ const SettingModal = () => {
     });
   };
 
+  function changeTheme() {
+    const newTheme = theme === 'light' ? 'dark' : 'light';
+    document.documentElement.classList.remove(theme);
+    document.documentElement.classList.add(newTheme);
+    dispatch(changeUserTheme(newTheme));
+    dispatch(updateUserThunk({theme:newTheme}))
+  }
+
   return (
     <SettingContainer>
       <TitleSwitcherWrap>
         <h1>{t('setting')}</h1>
+        <ThemeToggler onClick={changeTheme}></ThemeToggler>
       </TitleSwitcherWrap>
       <FormWrapper
         onSubmit={handleSubmit(onSubmit)}
